Fix MQTT password encryption
Fixed an issue introduced with the encryption of the MQTT password in the database that was causing both gateway.py and HA_HVAC_integration.py to use the password from the first row in the mqtt table.
Mqtt_passwd_decrypt.php has been modified to take an argument as input to specify if the password for the MQTT sensors account or the password for the HA  integration account should be retrieved from the mqtt table. Gateway.py and HA_HVAC_integration.py have been modified to call mqtt_passwd_decrypt.php with the correct argument.

A check has also been introduced in gateway.py to avoid opening multiple MQTT connections.